controllers players and coaches, models

+++ controllers/coachesController.js
@@ -60,12 +60,13 @@ export const coachesController = {
                 }
 
                 await fs.remove(req.files.image.tempFilePath)
-            } else {
-                image = {
-                    url: 'https://res.cloudinary.com/dlah9v2do/image/upload/v1684277158/userimage_wmdcqv.png',
-                    public_id: "FotosPerfil/qwloglssre2ryz5mgu9k"
-                }
-            }
+            } 
+            // else {
+            //     image = {
+            //         url: 'https://res.cloudinary.com/dlah9v2do/image/upload/v1684277158/userimage_wmdcqv.png',
+            //         public_id: "FotosPerfil/qwloglssre2ryz5mgu9k"
+            //     }
+            // }
 
             const newCoach = new Coach({
                 name,
@@ -141,7 +142,7 @@ export const coachesController = {
         try {
             const coachRemoved = await Coach.findById(req.params.id);
 
-            if (coachRemoved.image.public_id && coachRemoved.image.url !== 'https://res.cloudinary.com/dlah9v2do/image/upload/v1684277158/userimage_wmdcqv.png') {
+            if (coachRemoved.image !== null && coachRemoved.image.public_id && coachRemoved.image.url !== 'https://res.cloudinary.com/dlah9v2do/image/upload/v1684277158/userimage_wmdcqv.png') {
                 await deleteImage(coachRemoved.image.public_id)
             }
